pyproject.toml
[project]
name = "logic-extraction"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "lark>=1.1.0",
    "z3-solver>=4.12.0",
    "PyPDF2>=3.0.0",
    "python-dotenv>=1.0.0",
    "flask>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["parser*", "prover*", "domain*", "llm*", "grammar*", "web*"]
exclude = ["tests*", "extraction_output*"]
